Remove Freemium dependancy from shared module
