+++ macOS/LocalPackages/DataBrokerProtectionShared/Package.swift
@@ -31,7 +31,6 @@ let package = Package(
     ],
     dependencies: [
         .package(path: "../../BrowserServicesKit"),
-        .package(path: "../Freemium"),
         .package(path: "../NetworkProtectionMac"),
     ],
     targets: [
@@ -42,7 +41,6 @@ let package = Package(
                 .product(name: "PixelKit", package: "BrowserServicesKit"),
                 .product(name: "Configuration", package: "BrowserServicesKit"),
                 .product(name: "Persistence", package: "BrowserServicesKit"),
-                .product(name: "Freemium", package: "Freemium"),
                 .product(name: "NetworkProtectionIPC", package: "NetworkProtectionMac"),
             ],
             resources: [.copy("Resources")],
@@ -58,7 +56,6 @@ let package = Package(
                 .product(name: "PixelKit", package: "BrowserServicesKit"),
                 .product(name: "Configuration", package: "BrowserServicesKit"),
                 .product(name: "Persistence", package: "BrowserServicesKit"),
-                .product(name: "Freemium", package: "Freemium"),
             ],
             swiftSettings: [
                 .define("DEBUG", .when(configuration: .debug))
@@ -70,7 +67,6 @@ let package = Package(
                 "DataBrokerProtectionShared",
                 "DataBrokerProtectionSharedTestsUtils",
                 "BrowserServicesKit",
-                "Freemium",
                 .product(name: "PersistenceTestingUtils", package: "BrowserServicesKit"),
                 .product(name: "SubscriptionTestingUtilities", package: "BrowserServicesKit"),
             ],
